fix(ci): remove conflicting yosys installation
The 'formal-proof' job was installing the OSS CAD Suite, which
includes a version of Yosys that is not compatible with amaranth v0.4.
This was causing the 'read_ilang' command to be not found.

This commit removes the installation of the OSS CAD Suite to ensure that
the version of Yosys from the 'amaranth-yosys' package is used.